+++ CHANGELOG.md
@@ -9,7 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 ### Changed
+
 ### Fixed
+
+- Gracefully shut down origin server on `SIGTERM`.
+
 ### Removed
 
 ## [0.1.0] - 2024-11-05